+++ Sources/MapboxNavigation/PassiveLocationManager.swift
@@ -1,7 +1,5 @@
 import CoreLocation
-import UIKit
 import MapboxCoreNavigation
-import MapboxMaps
 
 /**
  An object that notifies a map view when the user’s location changes, minimizing the noise that normally accompanies location updates from a `CLLocationManager` object.
@@ -10,7 +8,7 @@ import MapboxMaps
  
  This class depends on `PassiveLocationDataSource` to detect the user’s location as it changes. If you want location updates but do not need to display them on a map and do not want a dependency on the MapboxNavigation module, you can use `PassiveLocationDataSource` instead of this class.
  */
-open class PassiveLocationManager: NSObject, LocationProvider {
+open class PassiveLocationManager: NavigationLocationProvider {
     /**
      Initializes the location manager with the given data source.
      
@@ -18,91 +16,21 @@ open class PassiveLocationManager: NSObject, LocationProvider {
      */
     public init(dataSource: PassiveLocationDataSource) {
         self.dataSource = dataSource
-        self.locationProviderOptions = LocationOptions()
         
         super.init()
         dataSource.delegate = self
+        self.locationManager = dataSource.systemLocationManager
     }
-    
-    /**
-     The location manager’s delegate.
-     */
-    public weak var delegate: LocationProviderDelegate?
     
     /**
      The location manager’s data source, which detects the user’s location as it changes.
      */
     public let dataSource: PassiveLocationDataSource
-
-    public var authorizationStatus: CLAuthorizationStatus {
-        CLLocationManager.authorizationStatus()
-    }
-
-    public func requestAlwaysAuthorization() {
-        dataSource.systemLocationManager.requestAlwaysAuthorization()
-    }
-
-    public func requestWhenInUseAuthorization() {
-        dataSource.systemLocationManager.requestWhenInUseAuthorization()
-    }
     
-    public var locationProviderOptions: LocationOptions
-    
-    public var accuracyAuthorization: CLAccuracyAuthorization {
-        if #available(iOS 14.0, *) {
-            return dataSource.systemLocationManager.accuracyAuthorization
-        } else {
-            return .fullAccuracy
-        }
-    }
-    
-    public var heading: CLHeading? {
-        return dataSource.systemLocationManager.heading
-    }
-    
-    // TODO: Consider replacing with public property.
-    public func setDelegate(_ delegate: LocationProviderDelegate) {
-        self.delegate = delegate
-    }
-    
-    @available(iOS 14.0, *)
-    public func requestTemporaryFullAccuracyAuthorization(withPurposeKey purposeKey: String) {
-        // CLLocationManager.requestTemporaryFullAccuracyAuthorization(withPurposeKey:) was introduced in the iOS 14 SDK in Xcode 12, so Xcode 11 doesn’t recognize it.
-        let requestTemporaryFullAccuracyAuthorization = Selector(("requestTemporaryFullAccuracyAuthorizationWithPurposeKey:" as NSString) as String)
-        guard dataSource.systemLocationManager.responds(to: requestTemporaryFullAccuracyAuthorization) else {
-            return
-        }
-        dataSource.systemLocationManager.perform(requestTemporaryFullAccuracyAuthorization, with: purposeKey)
-    }
-
-    public func startUpdatingLocation() {
+    public override func startUpdatingLocation() {
         dataSource.startUpdatingLocation()
     }
-
-    public func stopUpdatingLocation() {
-        dataSource.systemLocationManager.stopUpdatingLocation()
-    }
-
-    public var headingOrientation: CLDeviceOrientation {
-        get {
-            dataSource.systemLocationManager.headingOrientation
-        }
-        set {
-            dataSource.systemLocationManager.headingOrientation = newValue
-        }
-    }
-
-    public func startUpdatingHeading() {
-        dataSource.systemLocationManager.startUpdatingHeading()
-    }
-
-    public func stopUpdatingHeading() {
-        dataSource.systemLocationManager.stopUpdatingHeading()
-    }
-
-    public func dismissHeadingCalibrationDisplay() {
-        dataSource.systemLocationManager.dismissHeadingCalibrationDisplay()
-    }
+    
 }
 
 extension PassiveLocationManager: PassiveLocationDataSourceDelegate {